Move tests to dedicated test/ directory
Keep src/ as the shippable surface; tests + golden snapshots live in test/. typecheck stays scoped to src (the published code); tests are validated by execution (tsx) and linted (eslint src test).
