+++ test/snapshot.test.ts
@@ -11,7 +11,7 @@ import { readFileSync, writeFileSync, existsSync, mkdirSync } from 'node:fs'
 import { fileURLToPath } from 'node:url'
 import { dirname, join } from 'node:path'
 import JSZip from 'jszip'
-import pptxgen from './pptxgen'
+import pptxgen from '../src/pptxgen'
 
 const SNAP_DIR = join(dirname(fileURLToPath(import.meta.url)), '__snapshots__')
